Rename module to fix docs issues
